==> Code/1_Initial_Relative_Mortality_Analysis.ipynb <==
#Code written by Victoria Dunkley

# Understanding the Reach and Impact of the Centers for Disease Control and Prevention’s Women’s Health Research, 2018–2023
#### Code Notebook Goal: Analyze 2018-2021 National Vital Statistics System (NVSS) mortality data with a focus on discerning which conditions result in high relative mortality risk for US females. CDC-auhtored publications will be identified via systematic search for contitions with high relative mortality risk for US females
#### Initial Data pulled from :
   1. CDC Wonder Query for sex-stratified age-adjusted dataset (df_nvss): https://wonder.cdc.gov/controller/saved/D158/D450F314
   2. CDC Wonder Query for overarching sex-stratified 2018-2021 mortality (CDC_WONDER_TOTALS ): https://wonder.cdc.gov/controller/saved/D158/D402F797 ( this populates table 1)

#### In this notebook this analysis will:
1. Import NVSS mortality data from CDC Wonder
2. Restructure sex-stratified mortality data from long to wide format
3. Calculate relative mortality risk ( female age-adjusted death rate/male age-adjusted death rate) for US females and additional relevant mortality metrics
4. Export the dataset that will be made into a supplemental data table (CDC_WONDER_NCHS_MERGE)




## Load libraries and data

## import modules (base kernel is Python 3.11.7)
import numpy as np
import pandas as pd
import os
from IPython.display import Image
import matplotlib
import matplotlib.pyplot as plt
import seaborn as sns

## Enable multiple outputs from jupyter cells
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

## disable the Pandas "setting a copy of a slice" warning
pd.options.mode.chained_assignment = None

## set default number of DataFrame rows printed to 20
pd.set_option('display.max_rows', 20)

#get working directory
os.getcwd()

# Change working directory to one folder up
os.chdir('..')
os.getcwd()

# Import NVSS 2018-2021 mortality data and get the general shape of the data
df_nvss=pd.read_excel("Data/NVSS_2018_2021_Sex_Stratified_NCHS_Grouping_AgeAdjusted.xlsx") 
df_nvss.info()

## Import data with deaths per year per sex and rate info (overarching 2018-2021 mortality numbers)
CDC_WONDER_TOTALS = pd.read_excel("Data/CDC_WONDER_4_YEAR_TOTAL.xlsx")  
CDC_WONDER_TOTALS.head(20)

## Data Restructuring

# parse out 2018-2021 sex-stratified total deaths numbers
female_deaths = CDC_WONDER_TOTALS.loc[4,'Deaths']
male_deaths = CDC_WONDER_TOTALS.loc[9,'Deaths']

#Ensure unreliable and "" are NA
df_nvss= df_nvss.replace(['', 'Unreliable'], np.nan)

#check to make sure unreliable has been made NA.. should return empty
unreliable_rows = df_nvss.loc[(df_nvss['Crude Rate'] == 'Unreliable') | (df_nvss['Age Adjusted Rate'] == 'Unreliable')]


# check... should return empty rows
unreliable_rows

# create sex-stratified datasets... data is currently in long format and we want to restructure it to wide
df_female=df_nvss.loc[(df_nvss['Sex']=='Female')]
df_female.head(5)

df_male=df_nvss.loc[(df_nvss['Sex']=='Male')]
df_male.head(5)

#sort based on Cause of death Code prior to re-merging
df_female= df_female.sort_values(["ICD-10 113 Cause List"])

df_male= df_male.sort_values(["ICD-10 113 Cause List"])

# Remerge data to format data from long to wide
CDC_WONDER_NCHS_MERGE=\
    pd.merge(df_female,
             df_male,
             on=["ICD-10 113 Cause List"],
             how='outer',
            suffixes=('_F', '_M')) # designates what comes from df_female v. df_male

# view data shape and data presentation
CDC_WONDER_NCHS_MERGE.info()
CDC_WONDER_NCHS_MERGE.head()

CDC_WONDER_NCHS_MERGE['Population_F']=female_deaths
CDC_WONDER_NCHS_MERGE['Population_M']=male_deaths

## Analysis to discern metrics by sex and condition:
- Total deaths
- Relative Mortality Risk
- Mortality Burden


# sum male and female deaths for total deaths due to a specific cause
# if row is NA fill with 0 to get accurate numeric total death
CDC_WONDER_NCHS_MERGE['Total_deaths'] = CDC_WONDER_NCHS_MERGE['Deaths_F'].fillna(0) + CDC_WONDER_NCHS_MERGE['Deaths_M'].fillna(0)
CDC_WONDER_NCHS_MERGE.head()

# sort by total death to get top causes of death (aligns with literature)
CDC_WONDER_NCHS_MERGE.sort_values(['Total_deaths'], ascending=[False]).head(20)

#ensure age-adjusted death rates are numeric
CDC_WONDER_NCHS_MERGE["Age Adjusted Rate_F"] = pd.to_numeric(CDC_WONDER_NCHS_MERGE["Age Adjusted Rate_F"], errors="coerce")
CDC_WONDER_NCHS_MERGE["Age Adjusted Rate_M"] = pd.to_numeric(CDC_WONDER_NCHS_MERGE["Age Adjusted Rate_M"], errors="coerce")

#calculate relative mortality risk for females using age-adjusted death rates
CDC_WONDER_NCHS_MERGE["Relative Mortality Risk_F"] =CDC_WONDER_NCHS_MERGE["Age Adjusted Rate_F"]/CDC_WONDER_NCHS_MERGE["Age Adjusted Rate_M"]

#calculate relative mortality risk for males using age-adjusted death rates
CDC_WONDER_NCHS_MERGE["Relative Mortality Risk_M"]=CDC_WONDER_NCHS_MERGE["Age Adjusted Rate_M"]/CDC_WONDER_NCHS_MERGE["Age Adjusted Rate_F"]

#of deaths due to this cause in females/ total # of deaths in females ( Mortality burden)
CDC_WONDER_NCHS_MERGE["Female Mortality Burden"] = CDC_WONDER_NCHS_MERGE["Deaths_F"] /female_deaths

CDC_WONDER_NCHS_MERGE["Male Mortality Burden"]=CDC_WONDER_NCHS_MERGE["Deaths_M"] /male_deaths
CDC_WONDER_NCHS_MERGE

### Export final dataset
Export dataset to discern conditions with high relative mortality risk amongst females

# the following notebook will reformat  and better organize this information to generate a table 1
CDC_WONDER_NCHS_MERGE.to_excel("Results/RelativeMortalityInitialOutput.xlsx")
