==> Code/2_Creation_of_Supplemental_Table.ipynb <==
#Code written by Victoria Dunkley

# Understanding the Reach and Impact of the Centers for Disease Control and Prevention’s Women’s Health Research, 2018–2023
#### Code Notebook Goal: The goal of this notebook is to resturcture the dataset created in the previous code notebook (1_Initial_Realtive_Mortality_Analysis.ipynb) to create a supplemental data table. 
#### Restructing steps will include:
   1. Dropping non-specific conditions found in the the National Center for Health Statistics’ (NCHS) List of 113 Selected Causes of Death and Enterocolitis due to Clostridium difficile, and COVID-19 64 (Updated October 2020 to include WHO updates to ICD-10 for data year 2020). See Table B: https://www.cdc.gov/nchs/data/dvs/Part9InstructionManual2020-508.pdf
   2. Removing and renaming columns
   3. Parsing out relevant text data
   4. Fixing labels for text data irregularities
#### Data:
- Import: __RelativeMortalityInitialOutput.xlsx__
- Export: __SupplementalDataTable.xlsx__ 



## Load libraries and data

## import modules
import numpy as np
import pandas as pd
import os
from IPython.display import Image
import matplotlib
import matplotlib.pyplot as plt
import seaborn as sns

## Enable multiple outputs from jupyter cells
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

## disable the Pandas "setting a copy of a slice" warning
pd.options.mode.chained_assignment = None

## set default number of DataFrame rows printed to 20
pd.set_option('display.max_rows', 20)

#get working directory
os.getcwd()

# Change working directory to one folder up
#un comment line below this
os.chdir('..')
os.getcwd()

## Import data 
results = pd.read_excel("Results/RelativeMortalityInitialOutput.xlsx")  


#see initial state of the data
results.info()
results.head()

### Parse out non-relevant data


#Drop "overarching" conditions... these conditions are not apart of the 113 conditions we want to capture but are larger subset groupings.
values_to_drop = ['GR113-004', 'GR113-019', 'GR113-037', 'GR113-047','GR113-053','GR113-058','GR113-061','GR113-064','GR113-072',
                  'GR113-076','GR113-079','GR113-082','GR113-093','GR113-097','GR113-105','GR113-112','GR113-113',
                  'GR113-117','GR113-124','GR113-127',"GR113-131","GR113-054"]


# Dropping rows with specified values
results_filtered = results[~results['ICD-10 113 Cause List Code_F'].isin(values_to_drop)]

#check of 113 conditions
results_filtered.info()

### Delete non-relevant columns

# remove columns you don't need
columns_to_drop = ['Unnamed: 0', 'ICD-10 113 Cause List Code_F', 'Sex_F', 'Sex Code_F', 'ICD-10 113 Cause List Code_M', 'Sex_M', 'Sex Code_M',]

results_filtered.drop(columns=columns_to_drop, inplace=True)

results_filtered.info()

### Parse out ICD-10 code and condition name information

# Parse out ICD-10 code from disease name
results_filtered['ICD-10'] = results_filtered['ICD-10 113 Cause List'].str.extract(r'\((.*?)\)')[0]

# Removing the extracted content from the original column
results_filtered['ICD-10 113 Cause List'] = results_filtered['ICD-10 113 Cause List'].str.replace(r'\s*\(.*?\)', '', regex=True)

# Displaying the first 15 rows of the modified data
results_filtered.head(15)

# ensure that icd-10 column is a string
results_filtered['ICD-10'] = results_filtered['ICD-10'].astype('string')
results_filtered.info()

# sort by highest relative mortality risk
results_filtered = results_filtered.sort_values(by=['Relative Mortality Risk_F'],ascending=False)

#reset index and drop the previous index
results_filtered.reset_index(drop=True, inplace=True)
results_filtered.head(15)

#Some rows were irregular and do not parse out correctly when you use regex. This code block mannually inputs ICD codes that did not parse out correctly.
results_filtered.loc[15,'ICD-10']= "D65-E07,E15-E34,E65-F99,G04-G14,G23-G25,G31-H93,K00-K22,K29-K3l,K50-K66,K71-K72,K75-K76,K83-M99,N13.0-N13.5,N13.7-N13.9,N14,N15.0-N15.8-N15.9,N20-N23,N28-N39,N4l-N64,N80-N98"

results_filtered.loc[66,'ICD-10']='*U01.0-*U01.3,*U01.5-*U01.9,*U02,X85-X92,X96-Y09,Y87.1'

results_filtered.loc[76,'ICD-10']='*U03,X60-X71,X75-X84,Y87.0'

results_filtered.loc[80,'ICD-10']='B20-B24'

results_filtered.loc[86,'ICD-10']='*U01.4,X93-X95'

results_filtered.loc[87,'ICD-10']='X72-X74'

results_filtered.head(15)
results_filtered.tail(25)

# Remove '#' from the specified column
results_filtered['ICD-10 113 Cause List'] = results_filtered['ICD-10 113 Cause List'].str.replace('#', '', regex=False)

#check if # has been removed
filtered_df = results_filtered[results_filtered['ICD-10 113 Cause List'].str.contains('#')]

# should return an empty dataset
print(filtered_df)

### Final dataset restructuring

# Renaming columns
results_filtered.rename(columns={'ICD-10 113 Cause List': 'NCHS 113 Cause List', 'Deaths_F': 'Female Deaths', 'Population_F': 'Female Population', 'Deaths_M': 'Male Deaths', 'Population_M': 'Male Population'}, inplace=True)

results_filtered.info()



# Reordering columns
results_filtered = results_filtered[['NCHS 113 Cause List', 'ICD-10', 'Female Deaths','Female Population','Age Adjusted Rate_F','Relative Mortality Risk_F', 'Male Deaths','Male Population','Age Adjusted Rate_M','Relative Mortality Risk_M','Total_deaths','Female Mortality Burden','Male Mortality Burden']]
results_filtered.head()

### Export Supplemental Dataset

#export to excel so we can examine the thresholds for relative mortality ratio
results_filtered.to_excel("Results/SupplementalDataTable.xlsx")
