==> Code/3_Merge_Indicator_Information_to_SciClips.ipynb <==
#Code written by Victoria Dunkley

# Understanding the Reach and Impact of the Centers for Disease Control and Prevention’s Women’s Health Research, 2018–2023
#### Code Notebook Goal: The goal of this notbeook is to merge CDC-authored publications on conditions of interest to media, academic, and policy data queried from Altmetric and BMJ Impact Analytics.

##### **Conditions of Interest (identified using female relative mortality risk)**:
 - Alzheimer disease
 - Breast Cancer
 - Infections of the Kidney
 - Acute rheumatic fever and chronic rheumatic heart diseases

  **Bibliometric indicators for all CDC-authored publications published within the same timeframe will be calculated as a baseline comparision**

##### Data Sources Explained:
 1. __CDC Science Clips__: CDC-authored publications on codnitions of interest were identified by CDC librarians using CDC Science Clips (2018-2023).
    - Publications returned using systematic search terms were accessed from CDC Science Clips on September 24, 2024.      
 2. __Altmetric__: data pulled from Altemetric will be used to calulcate media attentnion, academic citation, and policy citation metrics.
    - Bibliometric indicators were calculated using data from publications indexed in Altmetric and BMJ Impact Analytics accessed on December 16, 2024.
 3. __BMJ Impact Analytics__:data pulled from BMJ will be used to calulcate the policy citation metric.
    - Bibliometric indicators were calculated using data from publications indexed in Altmetric and BMJ Impact Analytics accessed on December 16, 2024.


## import modules
import numpy as np
import pandas as pd
import os
from IPython.display import Image
import matplotlib
import matplotlib.pyplot as plt
import seaborn as sns

## Enable multiple outputs from jupyter cells
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

## disable the Pandas "setting a copy of a slice" warning
pd.options.mode.chained_assignment = None

## set default number of DataFrame rows printed to 20
pd.set_option('display.max_rows', 20)

# #get working directory
os.getcwd()
os.chdir('..')
os.getcwd()

### Science Clips quality check, cleaning, and combining

#### Data Import

# Science Clips data:
df_sciclips_total= pd.read_excel("Results/Bibliometrics/SciClips Publications/2018_2023_SciClips_Publications_SciClips.xlsx")
df_alz_total= pd.read_excel("Results/Bibliometrics/SciClips Publications/2018_2023_Alzheimer_Publications_SciClips.xlsx")
df_breastcancer_total= pd.read_excel("Results/Bibliometrics/SciClips Publications/2018_2023_BreastCancer_Publications_SciClips.xlsx")
df_kidney_total= pd.read_excel("Results/Bibliometrics/SciClips Publications/2018_2023_Kidney_Publications_SciClips.xlsx")
df_rhu_total= pd.read_excel("Results/Bibliometrics/SciClips Publications/2018_2023_Rheumatic_Publications_SciClips.xlsx")

# Altmetric data:
df_sciclips_alt= pd.read_excel("Results/Bibliometrics/Altmetric Data Pull/Altmetric_20241216_2018_2023_SciClips_Publications_SciClips.xlsx")
df_alz_alt= pd.read_excel("Results/Bibliometrics/Altmetric Data Pull/Altmetric_20241216_2018_2023_Alzheimers_Publications_SciClips.xlsx")
df_breastcancer_alt= pd.read_excel("Results/Bibliometrics/Altmetric Data Pull/Altmetric_20241216_2018_2023_BreastCancer_Publications_SciClips.xlsx")
df_kidney_alt= pd.read_excel("Results/Bibliometrics/Altmetric Data Pull/Altmetric_20241216_2018_2023_Kidney_Publications_SciClips.xlsx")
df_rhu_alt= pd.read_excel("Results/Bibliometrics/Altmetric Data Pull/Altmetric_20241216_2018_2023_Rheumatic_Publications_SciClips.xlsx")

#BMJ
df_sciclips_bmj= pd.read_excel("Results/Bibliometrics/BMJ Data Pull/BMJ_20241216_2018_2020_SciClips_Publications_SciClips.xlsx")
df_alz_bmj= pd.read_excel("Results/Bibliometrics/BMJ Data Pull/BMJ_20241216_2018_2020_Alzheimers_Publications_SciClips.xlsx")
df_breastcancer_bmj= pd.read_excel("Results/Bibliometrics/BMJ Data Pull/BMJ_20241216_2018_2020_BreastCancer_Publications_SciClips.xlsx")
df_kidney_bmj= pd.read_excel("Results/Bibliometrics/BMJ Data Pull/BMJ_20241216_2018_2020_Kidney_Publications_SciClips.xlsx")
df_rhu_bmj= pd.read_excel("Results/Bibliometrics/BMJ Data Pull/BMJ_20241216_2018_2020_Rheumatic_Publications_SciClips.xlsx")

#### Merge, Filter, and restructure Science Clips and bibliometric indicator data

#Examine shape
df_sciclips_total.info()

#Examine shape
df_sciclips_alt.info()

df_sciclips_alt.head()

#keep all but this row
df_sciclips_alt = df_sciclips_alt[df_sciclips_alt['Details Page URL'] != 'https://www.altmetric.com/details/72325352']

#ensure all DOI are lowercase prior to merge
df_sciclips_bmj['DOI'] = df_sciclips_bmj['DOI'].str.lower()
df_sciclips_total['DOI'] = df_sciclips_total['DOI'].str.lower()
df_sciclips_alt['DOI'] = df_sciclips_alt['DOI'].str.lower()

#merge in altmetric info
df_clips_DOI = pd.merge(df_sciclips_total,df_sciclips_alt,how='left',on='DOI')

#get post merge dataset shape
df_clips_DOI.info()

#filter to only keep rows that are not NA
df_clips_DOI = df_clips_DOI[~df_clips_DOI['Details Page URL'].isna()]

df_clips_DOI.info()

#only the rows where the 'DOI' is not present
df_sciclips_total = df_sciclips_total[~df_sciclips_total['DOI'].isin(df_clips_DOI['DOI'])]

df_sciclips_total.info()

 #only the rows where the 'Details Page URL' is not present
df_sciclips_alt = df_sciclips_alt[~df_sciclips_alt['Details Page URL'].isin(df_clips_DOI['Details Page URL'])]

df_sciclips_alt.info()

#rename for easier merge
df_sciclips_alt.rename(columns={'PubMed ID': 'PMID'}, inplace=True)

#PMID merge
df_clips_PMID = pd.merge(df_sciclips_total,df_sciclips_alt,how = 'left',on='PMID')

df_clips_PMID.info()

#keep is not NA rows
df_clips_PMID = df_clips_PMID[~df_clips_PMID['Details Page URL'].isna()]

#we will concat using this later
df_sciclips_total = df_sciclips_total[~df_sciclips_total['PMID'].isin(df_clips_PMID['PMID'])]

df_clips_DOI.info()

df_clips_DOI.rename(columns={'Title_x': 'Title'}, inplace=True)

df_clips_PMID.info()

remove_columns_PMID = df_sciclips_alt.columns.to_list()
remove_columns_PMID.remove('Altmetric Attention Score')
remove_columns_PMID.remove('DOI')
remove_columns_PMID.remove('Title')
remove_columns_PMID.remove('PMID')
remove_columns_PMID.remove('Policy mentions')
remove_columns_PMID.remove('Number of Dimensions citations')
remove_columns_PMID.extend(['Title_y', 'DOI_y'])

df_clips_PMID.drop(remove_columns_PMID,axis=1,inplace=True)
df_clips_PMID.info()

df_clips_PMID.rename(columns={'Title_x': 'Title', 'DOI_x': 'DOI'}, inplace=True)

#concat to get a full view of sciclips
df_clips = pd.concat([df_clips_DOI,df_clips_PMID,df_sciclips_total])

df_clips.info()

#drop BMJ columns that are not of interest or already in the original dataset
columns_to_drop = ['Title', 'Journal','Authors','ORCIDs','Your tags']  # Specify the columns you want to drop
df_sciclips_bmj = df_sciclips_bmj.drop(columns=columns_to_drop)
df_sciclips_bmj.info()

#merge in bmj info
df_clips = pd.merge(df_clips,df_sciclips_bmj,how = 'left',on='DOI')
df_clips.info()
df_clips.head()

#Ensure numeric columns that will be used to calculate indicators have 0 in place of NA
df_clips['Altmetric Attention Score'] = df_clips['Altmetric Attention Score'].fillna(0)
df_clips['Number of Dimensions citations'] = df_clips['Number of Dimensions citations'].fillna(0)
df_clips['Policy mentions'] = df_clips['Policy mentions'].fillna(0)
df_clips['Policy citation count'] = df_clips['Policy citation count'].fillna(0)

# add an any policy flag column
df_clips['Any Policy'] = ((df_clips['Policy mentions'] > 0) | (df_clips['Policy citation count'] > 0)).astype(int)
df_clips.info()
df_clips.head()

#Putting the final columns in the final df
df_clips = df_clips.iloc[:, [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 
                              19,46,64,65,66,67,68,69,72]] 
                           
df_clips.info()

#export pubs of interest and indicators
df_clips.to_excel("Results/Bibliometrics/SciClips Publication plus Indicators/20250116_2018_2023_SciClips_Publications_SciClips_Indicators.xlsx")

## Conditions of Interest
CDC-auhtored publications will be merged to bibliometric data from Altermetric and BMJ Impact Analytics.

#### Alzheimer's disease

#look at all datasets pertaining to this condition
df_alz_total.info()
df_alz_alt.info()
df_alz_bmj.info()

#merge pubs and altmetric info
df_alz = pd.merge(df_alz_total,df_alz_alt,how='left',on='DOI')
df_alz.info()

#drop BMJ columns that are not of interest or already in the original dataset
columns_to_drop = ['Title', 'Journal','Authors','ORCIDs','Your tags']  
df_alz_bmj = df_alz_bmj.drop(columns=columns_to_drop)
df_alz_bmj.info()

#merge in bmj info
df_alz = pd.merge(df_alz,df_alz_bmj,how='left',on='DOI')
df_alz.info()

# fill numeric variables that will be used to calculate indicators with 0
df_alz['Altmetric Attention Score'] = df_alz['Altmetric Attention Score'].fillna(0)
df_alz['Number of Dimensions citations'] = df_alz['Number of Dimensions citations'].fillna(0)
df_alz['Policy mentions'] = df_alz['Policy mentions'].fillna(0)
df_alz['Policy citation count'] = df_alz['Policy citation count'].fillna(0)

# add an any policy flag column
df_alz['Any Policy'] = ((df_alz['Policy mentions'] > 0) | (df_alz['Policy citation count'] > 0)).astype(int)
df_alz.info()
df_alz.head()

#only keep relevant columns
df_alz = df_alz.iloc[:, [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 
                              19,46,64,65,66,67,68,69,72]] 
                           
df_alz.info()

#export pubs of interest and indicators

df_alz.to_excel("Results/Bibliometrics/SciClips Publication plus Indicators/20250116_2018_2023_Alzheimer_Publications_SciClips_Indicators.xlsx")

### Breast cancer

#look at all datasets pertaining to this condition
df_breastcancer_total.info()
df_breastcancer_alt.info()
df_breastcancer_bmj.info()

#merge in the altmetric info
df_breastcancer = pd.merge(df_breastcancer_total,df_breastcancer_alt,how='left',on='DOI')
df_breastcancer.info()

#drop BMJ columns that are not of interest or already in the original dataset
columns_to_drop = ['Title', 'Journal','Authors','ORCIDs','Your tags']  # Specify the columns you want to drop
df_breastcancer_bmj = df_breastcancer_bmj.drop(columns=columns_to_drop)
df_breastcancer_bmj.info()

#merge in bmj
df_breastcancer = pd.merge(df_breastcancer,df_breastcancer_bmj,how='left',on='DOI')
df_breastcancer.info()

# fill numeric variables that will be used to calculate indicators with 0
df_breastcancer['Altmetric Attention Score'] = df_breastcancer['Altmetric Attention Score'].fillna(0)
df_breastcancer['Number of Dimensions citations'] = df_breastcancer['Number of Dimensions citations'].fillna(0)
df_breastcancer['Policy mentions'] = df_breastcancer['Policy mentions'].fillna(0)
df_breastcancer['Policy citation count'] = df_breastcancer['Policy citation count'].fillna(0)

# add an any policy flag column
df_breastcancer['Any Policy'] = ((df_breastcancer['Policy mentions'] > 0) | (df_breastcancer['Policy citation count'] > 0)).astype(int)
df_breastcancer.info()
df_breastcancer.head()

#only keep relevant columns
df_breastcancer = df_breastcancer.iloc[:, [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 
                              19,46,64,65,66,67,68,69,72]] 
                           
df_breastcancer.info()

#export pubs of interest and indicators
df_breastcancer.to_excel("Results/Bibliometrics/SciClips Publication plus Indicators/20250116_2018_2023_BreastCancer_Publications_SciClips_Indicators.xlsx")

### Kidney

df_kidney_total.info()
df_kidney_alt.info()
df_kidney_bmj.info()

#merge in altmetric
df_kidney = pd.merge(df_kidney_total,df_kidney_alt,how='left',on='DOI')
df_kidney.info()

#drop BMJ columns that are not of interest or already in the original dataset
columns_to_drop = ['Title', 'Journal','Authors','ORCIDs','Your tags']  # Specify the columns you want to drop
df_kidney_bmj = df_kidney_bmj.drop(columns=columns_to_drop)
df_kidney_bmj.info()

#merge in bmj
df_kidney = pd.merge(df_kidney,df_kidney_bmj,how='left',on='DOI')
df_kidney.info()

# fill numeric variables that will be used to calculate indicators with 0
df_kidney['Altmetric Attention Score'] = df_kidney['Altmetric Attention Score'].fillna(0)
df_kidney['Number of Dimensions citations'] = df_kidney['Number of Dimensions citations'].fillna(0)
df_kidney['Policy mentions'] = df_kidney['Policy mentions'].fillna(0)
df_kidney['Policy citation count'] = df_kidney['Policy citation count'].fillna(0)

df_kidney['Any Policy'] = ((df_kidney['Policy mentions'] > 0) | (df_kidney['Policy citation count'] > 0)).astype(int)
df_kidney.info()
df_kidney.head()

#only keep relevant columns
df_kidney = df_kidney.iloc[:, [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 
                              19,46,64,65,66,67,68,69,72]] 
                           
df_kidney.info()

#export pubs of interest and indicators
df_kidney.to_excel("Results/Bibliometrics/SciClips Publication plus Indicators/20250116_2018_2023_Kidney_Publications_SciClips_Indicators.xlsx")

### Acute rheumatic fever and chronic rheumatic heart diseases

df_rhu_total.info()
df_rhu_alt.info()
df_rhu_bmj.info()

# merge in altmetric
df_rhu = pd.merge(df_rhu_total,df_rhu_alt,how='left',on='DOI')
df_rhu.info()

#drop BMJ columns that are not of interest or already in the original dataset
columns_to_drop = ['Title', 'Journal','Authors','ORCIDs','Your tags']  # Specify the columns you want to drop
df_rhu_bmj = df_rhu_bmj.drop(columns=columns_to_drop)
df_rhu_bmj.info()

#merge in bmj
df_rhu = pd.merge(df_rhu,df_rhu_bmj,how='left',on='DOI')
df_rhu.info()

# fill numeric variables that will be used to calculate indicators with 0
df_rhu['Altmetric Attention Score'] = df_rhu['Altmetric Attention Score'].fillna(0)
df_rhu['Number of Dimensions citations'] = df_rhu['Number of Dimensions citations'].fillna(0)
df_rhu['Policy mentions'] = df_rhu['Policy mentions'].fillna(0)
df_rhu['Policy citation count'] = df_rhu['Policy citation count'].fillna(0)

#create policy flag
df_rhu['Any Policy'] = ((df_rhu['Policy mentions'] > 0) | (df_rhu['Policy citation count'] > 0)).astype(int)
df_rhu.info()
df_rhu.head()

#only keep relevant columns
df_rhu = df_rhu.iloc[:, [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 
                              19,46,64,65,66,67,68,69,72]] 
                           
df_rhu.info()

#export pubs of interest and indicators
df_rhu.to_excel("Results/Bibliometrics/SciClips Publication plus Indicators/20250116_2018_2023_Rheumatic_Publications_SciClips_Indicators.xlsx")
